Fixed some bugs with the stat classes
Aggregate and Death were converting string to int when the value variable
is a long
Player class was not assigning a category through the constructor

+++ src/com/github/etsai/kfsxtrackingserver/stats/Aggregate.java
@@ -15,7 +15,7 @@ public class Aggregate extends Death {
     public static Aggregate build(ResultSet rs) throws SQLException {
         Aggregate aggregate= new Aggregate(rs.getInt("id"), rs.getString("stat"), 
                 rs.getString("category"));
-        aggregate.addValue(Integer.valueOf(rs.getString("value")));
+        aggregate.addValue(Long.valueOf(rs.getString("value")));
         return aggregate;
     }
     private final String category;
